[#293] рефакторинг, BluetoothActivity перенесена в отдельный пакет
