+++ SportiduinoManager/app/src/main/AndroidManifest.xml
@@ -33,7 +33,7 @@
             </intent-filter>
         </activity>
         <activity android:name="DatabaseActivity" />
-        <activity android:name="BluetoothActivity" />
+        <activity android:name=".activity.devices.BluetoothActivity" />
         <activity android:name="ChipInitActivity" />
     </application>
 </manifest>
