+++ SportiduinoManager/app/src/main/java/ru/mmb/sportiduinomanager/activity/devices/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Activity for bluetooth stations management.
+ */
+package ru.mmb.sportiduinomanager.activity.devices;
