+++ SportiduinoManager/app/src/main/java/ru/mmb/sportiduinomanager/activity/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Package for activities.
+ * Main activity parent will be placed here. Each descendant activity will have it's own package.
+ */
+package ru.mmb.sportiduinomanager.activity;
